+++ CHANGELOG.md
@@ -5,6 +5,7 @@ Yii2 EAuth Change Log
 * Fixed error param names for Facebook (#63)
 * Use the latest Graph Api v2.5 for Facebook (#65)
 * Fixed `makeRequest` method (#68)
+* Added `makeRequest` method to OpenID services
 
 ### 2.3.0 (17.10.2015)
 * Added InstagramOAuth2Service (#61)